fixup: Disassociated Anyp::Host from URI
... as the first step towards reducing AnyP::Host/Ssl::GeneralName
code duplication.

Also named Anyp::Host "domain name" variant type instead of using
generic SBuf. Eventually, DomainName may even become a class, but even
this small step already allows us to better document its type
requirements and APIs.

Also placed Anyp::Host::<Ip::Address> variant first because it is
simpler and because that variant order matches ip(),domainName() method
order.

Also moved Anyp::Host into dedicated files.
