+++ src/acl/ServerName.cc
@@ -11,6 +11,7 @@
 #include "squid.h"
 #include "acl/FilledChecklist.h"
 #include "acl/ServerName.h"
+#include "anyp/Host.h"
 #include "client_side.h"
 #include "http/Stream.h"
 #include "HttpRequest.h"
@@ -110,7 +111,7 @@ Acl::ServerNameCheck::match(ACLChecklist * const ch)
             // ASCII encoding"; "Literal IPv4 and IPv6 addresses are not
             // permitted". TODO: Store TlsDetails::serverName and similar
             // domains using a new domain-only type instead of SBuf.
-            clientRequestedServerName = AnyP::Host::FromDecodedDomain(clientSni);
+            clientRequestedServerName = AnyP::Host::FromSimpleDomainName(clientSni);
         }
 
         if (useConsensus) {
